+++ spec/yabeda/resque_spec.rb
@@ -8,6 +8,7 @@ RSpec.describe Yabeda::Resque do
 
     example.run
 
+    Timecop.unfreeze
     Resque.inline = original_inline
   end
 
@@ -44,8 +45,9 @@ RSpec.describe Yabeda::Resque do
 
   context "when job is processed" do
     it "increments successful job counter" do
-      Resque.inline = true
-      Resque.enqueue(DefaultJob)
+      allow(::Resque).to receive(:info).and_return({
+        processed: 1
+      })
 
       expect { Yabeda.collect! }.to \
         update_yabeda_gauge(Yabeda.resque.jobs_processed)
@@ -53,10 +55,99 @@ RSpec.describe Yabeda::Resque do
     end
   end
 
+  context "when a job is being worked on" do
+    let(:start_time) { Time.now.utc }
+    let(:queue) { "default" }
+    let(:working_workers) do
+      [
+        Resque::Worker.new(queue).tap { |w| w.job = {"queue" => "default", "run_at" => start_time.iso8601, "payload" => []} },
+        Resque::Worker.new(queue).tap { |w| w.instance_variable_set :@job, {"queue" => "default", "run_at" => (start_time - 60).iso8601, "payload" => []} },
+        Resque::Worker.new(queue).tap { |w| w.instance_variable_set :@job, {"queue" => "default", "run_at" => (start_time - 75).iso8601, "payload" => []} }
+      ]
+    end
+
+    before(:each) do
+      allow(::Resque).to receive(:working).and_return(working_workers)
+    end
+
+    context "when configured to measure in seconds" do
+      before(:each) do
+        Yabeda::Resque.install!(jobs_processing_oldest_age_unit: :seconds)
+      end
+
+      it "increments the jobs_processing_oldest_age" do
+        Timecop.freeze(start_time)
+        expect { Yabeda.collect! }.to \
+          update_yabeda_gauge(Yabeda.resque.jobs_processing_oldest_age)
+          .with(75)
+
+        Timecop.travel(15)
+        expect { Yabeda.collect! }.to \
+          update_yabeda_gauge(Yabeda.resque.jobs_processing_oldest_age)
+          .with(90)
+      end
+    end
+
+    context "when configured to measure in minutes" do
+      before(:each) do
+        Yabeda::Resque.install!(jobs_processing_oldest_age_unit: :minutes)
+      end
+
+      it "increments the jobs_processing_oldest_age" do
+        Timecop.freeze(start_time)
+        expect { Yabeda.collect! }.to \
+          update_yabeda_gauge(Yabeda.resque.jobs_processing_oldest_age)
+          .with(1.25)
+
+        Timecop.travel(15)
+        expect { Yabeda.collect! }.to \
+          update_yabeda_gauge(Yabeda.resque.jobs_processing_oldest_age)
+          .with(1.5)
+      end
+    end
+
+    context "when configured to measure in hours" do
+      before(:each) do
+        Yabeda::Resque.install!(jobs_processing_oldest_age_unit: :hours)
+      end
+
+      it "increments the jobs_processing_oldest_age" do
+        Timecop.freeze(start_time + 3600 - 75)
+        expect { Yabeda.collect! }.to \
+          update_yabeda_gauge(Yabeda.resque.jobs_processing_oldest_age).with(1)
+
+        Timecop.travel(3600 * 1.5)
+        expect { Yabeda.collect! }.to \
+          update_yabeda_gauge(Yabeda.resque.jobs_processing_oldest_age)
+          .with(2.5)
+      end
+    end
+
+    context "when configured to measure in days" do
+      before(:each) do
+        Yabeda::Resque.install!(jobs_processing_oldest_age_unit: :days)
+      end
+
+      it "increments the jobs_processing_oldest_age" do
+        # 1 day later, but ignore the 75 seconds to get clean values
+        Timecop.freeze(Time.at(start_time + (24 * 60 * 60) - 75))
+        expect { Yabeda.collect! }.to \
+          update_yabeda_gauge(Yabeda.resque.jobs_processing_oldest_age)
+          .with(1)
+
+        Timecop.travel(24 * 60 * 60 * 1.5)
+        expect { Yabeda.collect! }.to \
+          update_yabeda_gauge(Yabeda.resque.jobs_processing_oldest_age)
+          .with(2.5)
+      end
+    end
+  end
+
   context "when a job fails" do
     it "increments failed job counter" do
-      Resque.inline = true
-      Resque.enqueue(DefaultJob)
+      allow(::Resque).to receive(:info).and_return({
+        failed: 1
+      })
 
       expect { Yabeda.collect! }.to \
         update_yabeda_gauge(Yabeda.resque.jobs_failed)
@@ -75,8 +166,10 @@ RSpec.describe Yabeda::Resque do
 
   context "workers" do
     it "collects workers count" do
-      Resque.inline = true
-      Resque.enqueue(DefaultJob)
+      allow(::Resque).to receive(:info).and_return({
+        workers: 1,
+        working: 0
+      })
 
       expect { Yabeda.collect! }.to \
         update_yabeda_gauge(Yabeda.resque.workers_total)