+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## [Unreleased]
+- Add jobs_processing_oldest_age metric with unit as config option
 
 ## [1.1.0] - 2025-02-4
 - Fix: queue_sizes metric not being exported